requirements.txt
opencv-python
pytesseract
sympy
openai
matplotlib
streamlit
Pillow
